Consolidated Credentials Handling in SalesforceCredentials
Consolidated the credentials handling and added additional steps to retrieve the password + security token from the encrypted file on each login. Reduced number of times having to enter in password and passing the password around to different objects. Once retrieved and login session key is returned, the password is flushed from memory and does not get passed between objects.