Additional Asynchronous Processing on Profiles
Continuing to work on Profiles Asynchronous processing in the tool. Permission Sets are an additional factor which needs to be done as well. Profiles/Permission Sets are arduous and continue to plague the metadata retrieval process. Not sure if Salesforce will make additional future enhancements or not, but they are painful. May opt to retrieve Profiles/PermSets in another manner. Even tried Tooling and it did not provide any of the permissions against Objects, Tabs, etc. Minimizing the number of profiles is the only way to avoid timeout issues, but this means also requesting to retrieve all related items for a complete profile picture.