Converted Salesforce Credentials to Object Instance Class
Original class had static methods which handled the credentials. This was problematic when there was a login timeout issue as I had to handle each one individually. The object instance class is better suited for passing that as an object to other related forms which need to use the same credentials. Also cleaned up some legacy code and tightened up the build package xml in the DevelopmentEnvironment and GenerateDeploymentPackage