Adjusted Metadata retrieval for Activity
Adjusted the metadata retrieval to add the Activity object directly as this object is not considered an an actual object by Salesforce and won't add it dynamically, but contains all custom fields related to  the Task and Event object. However, the custom fields do not appear in Event or Task but in Activity, so have to add it manually for the retrieval of the custom fields.